// MVCApi.AngularClient/src/api/model/customerDto.ts
/**
 * MVCApi
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ContactInfoDto } from './contactInfoDto';
import { AddressDto } from './addressDto';


export interface CustomerDto { 
    id?: string;
    firstName?: string | null;
    lastName?: string | null;
    dateOfBirth?: string;
    addresses?: Array<AddressDto> | null;
    contactInfos?: Array<ContactInfoDto> | null;
}